Use same link/bg colors for calendar events in admin UI
